Move types and constants
