+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const MAGIC_PROXY_SYMBOL = Symbol('Magic proxy symbol');
